requirements.txt
numpy
pillow
matplotlib
scipy
opencv-python
torch
torchvision
